+++ bistro/orders/views.py
@@ -3,8 +3,6 @@ from decimal import Decimal
 
 from django.contrib import messages
 from django.core.exceptions import ObjectDoesNotExist
-from django.db import IntegrityError
-from django.db import transaction
 from django.db.models import F
 from django.db.models import Q
 from django.db.models import Sum
@@ -18,8 +16,8 @@ from .forms import OrderForm
 from .forms import SearchOrderForm
 from .forms import TotalPaidOrdersForm
 from .forms import UpdateStatusForm
+from .handlers.add_order import handle_add_order
 from .models import Order
-from .models import OrderItem
 
 
 def order_management_view(request):
@@ -48,42 +46,6 @@ def order_management_view(request):
         "modals": get_modals(),
     }
     return render(request, "orders/order_management.html", context)
-
-
-def handle_add_order(request):
-    form = OrderForm(request.POST)
-    if form.is_valid():
-        try:
-            with transaction.atomic():
-                order = form.save(commit=False)
-                order.save()
-
-                items = form.cleaned_data["items"]
-                quantities = form.cleaned_data["quantities"]
-
-                order_items = [
-                    OrderItem(order=order, menu_item=item, quantity=quantities[item.id])
-                    for item in items
-                ]
-                OrderItem.objects.bulk_create(order_items)
-                messages.success(
-                    request,
-                    f"Order for table {form.cleaned_data['table_number']} "
-                    f"added successfully.",
-                )
-                return redirect("orders:manager")
-
-        except IntegrityError as e:
-            messages.error(request, f"An error occurred while saving the order: {e!s}")
-        except Exception as e:  # noqa: BLE001
-            messages.error(request, f"An unexpected error occurred: {e!s}")
-
-    else:
-        messages.error(
-            request,
-            "Invalid form data. Please check the input and try again.",
-        )
-    return None
 
 
 def handle_delete_order(request):
